# docs/openapi.yml
openapi: 3.0.3
info:
  title: LilyOpenCMS API
  description: |
    A comprehensive API for LilyOpenCMS content management system.
    
    ## Features
    - **Content Management**: News, albums, chapters, and media
    - **User Management**: Profiles, authentication, and permissions
    - **Public API**: Multiplatform access for mobile apps and third-party integrations
    - **SEO Management**: Comprehensive SEO tools and analytics
    - **Analytics**: Performance monitoring and user statistics
    
    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Public API endpoints (`/api/public/*`) do not require authentication.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Please implement appropriate retry logic.
    
    ## Support
    For API support, please refer to the documentation or contact the development team.
  version: 1.0.0
  contact:
    name: LilyOpenCMS Development Team
    email: support@lilyopencms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}"
    description: Custom server (set scheme/host below to try any deployment)
    variables:
      scheme:
        enum:
          - http
          - https
        default: https
      host:
        default: localhost:5000
        description: Hostname and optional port (e.g., api.example.com or localhost:5000)
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/images:
    get:
      tags:
        - Media
      summary: List images with pagination, filters, and secure ownership-based access control
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: visibility
          in: query
          description: Filter by visibility (visible/hidden)
          schema: { type: string }
        - name: all_users
          in: query
          description: When true, applies proper ownership filtering - regular users see own images + admin/suadmin visible images, editors see own + assigned writers' + admin/suadmin visible images, admins see all images
          schema: { type: boolean }
      responses:
        '200': { description: Images list returned }
    post:
      tags:
        - Media
      summary: Upload a new image
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                is_visible: { type: boolean }
      responses:
        '201': { description: Image uploaded }
  /api/images/{image_id}:
    get:
      tags:
        - Media
      summary: Get image details
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Image details returned }
        '404': { description: Not found }
    put:
      tags:
        - Media
      summary: Update image
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        '200': { description: Image updated }
        '404': { description: Not found }
    delete:
      tags:
        - Media
      summary: Delete image
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
  /api/images/{image_id}/visibility:
    patch:
      tags:
        - Media
      summary: Toggle image visibility
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_visible: { type: boolean }
              required: [is_visible]
      responses:
        '200': { description: Visibility updated }
        '404': { description: Not found }
  /api/images/bulk-delete:
    post:
      tags:
        - Media
      summary: Bulk delete images
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: integer }
              required: [ids]
      responses:
        '200': { description: Bulk delete completed }
  /api/images/bulk-visibility:
    post:
      tags:
        - Media
      summary: Bulk update image visibility
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: integer }
                is_visible: { type: boolean }
              required: [ids, is_visible]
      responses:
        '200': { description: Bulk visibility updated }
  /api/images/{image_id}/usage:
    get:
      tags:
        - Media
      summary: Get image usage statistics
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Image usage returned }
        '404': { description: Not found }

  # =============================================================================
  # RATINGS ENDPOINTS
  # =============================================================================

  # =============================================================================
  # COMMENTS ENDPOINTS
  # =============================================================================
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via cookies
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external ad access and third-party integrations
security:
  - sessionAuth: []
  - bearerAuth: []